support wasm and js precompiled tests
